% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{br411}
\alias{br411}
\title{3d topological surface measurements for one land of a bullet from the Hamby study}
\format{an x3p file (a list)}
\usage{
br411
}
\description{
The scan shows land 1 of bullet 1 from barrel 4 of Hamby Set 252. The original scan is available from NRBTD.
}
\keyword{datasets}
