% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooves.R
\name{check_grooves}
\alias{check_grooves}
\title{Check grooves for correctness}
\usage{
check_grooves(x)
}
\arguments{
\item{x}{output from cc_locate_grooves}
}
\value{
TRUE if ok, error otherwise
}
\description{
Check grooves for correctness
}
