% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_bcp.R
\name{get_grooves_bcp}
\alias{get_grooves_bcp}
\title{Use Bayesian Changepoint Detection to identify groove locations}
\usage{
get_grooves_bcp(x, value, adjust = 10, return_plot = FALSE)
}
\arguments{
\item{x}{numeric vector of locations (in microns)}

\item{value}{numeric values of surface measurements in microns}

\item{adjust}{positive number to adjust the grooves}

\item{return_plot}{return plot of grooves?}
}
\description{
Use Bayesian Changepoint Detection to identify groove locations
}
