% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_lasso.R
\name{get_grooves_lasso}
\alias{get_grooves_lasso}
\title{Use logistic model to identify groove locations}
\usage{
get_grooves_lasso(
  x,
  value,
  lasso_method = "basic",
  pred_cutoff = ifelse(lasso_method == "basic", 0.3, 0.34),
  return_plot = F
)
}
\arguments{
\item{x}{numeric vector of locations (in microns)}

\item{value}{numeric values of surface measurements in microns}

\item{lasso_method}{use the 'basic' model or the 'full' model with interaction terms?}

\item{pred_cutoff}{equal error rate cutoff for classification into GEA or LEA, trained on Hamby set 44}

\item{return_plot}{return plot of grooves?}
}
\description{
Use logistic model to identify groove locations
}
