% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_logistic.R
\name{get_grooves_logisticlegacy}
\alias{get_grooves_logisticlegacy}
\title{Use logistic model to identify groove locations}
\usage{
get_grooves_logisticlegacy(x, value, adjust = 10, return_plot = F)
}
\arguments{
\item{x}{numeric vector of locations (in microns)}

\item{value}{numeric values of surface measurements in microns}

\item{adjust}{positive number to adjust the grooves}

\item{return_plot}{return plot of grooves?}
}
\description{
Use logistic model to identify groove locations
}
