% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_quadratic.R
\name{get_grooves_quadratic}
\alias{get_grooves_quadratic}
\title{Quadratic fit to find groove locations}
\usage{
get_grooves_quadratic(x, value, adjust, return_plot = F)
}
\arguments{
\item{x}{numeric vector of locations (in microns)}

\item{value}{numeric values of surface measurements in microns}

\item{adjust}{positive number to adjust the grooves}

\item{return_plot}{return plot of grooves?}
}
\value{
list of groove vector and plot of crosscut with shoulder locations
}
\description{
Use a robust fit of a quadratic curve to find groove locations
}
\examples{
\dontrun{
# Set the data up to be read in, cleaned, etc.
library(bulletxtrctr)
library(x3ptools)

example_data <- bullet_pipeline(
  location = list(Bullet1 = c(hamby252demo$bullet1[3])),
  stop_at_step = "crosscut",
  x3p_clean = function(x) x \%>\%
      x3p_scale_unit(scale_by = 10^6) \%>\%
      rotate_x3p(angle = -90) \%>\%
      y_flip_x3p()
)

get_grooves_quadratic(example_data$ccdata[[1]]$x,
  example_data$ccdata[[1]]$value,
  adjust = 30, return_plot = T
)
cc_locate_grooves(example_data$ccdata[[1]],
  method = "quadratic",
  adjust = 30, return_plot = T
)
}
}
