% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mask_hough.R
\name{get_mask_hough}
\alias{get_mask_hough}
\title{Create mask layer for grooves identified with hough}
\usage{
get_mask_hough(land.x3p, grooves)
}
\arguments{
\item{land.x3p}{an x3p file containing a bullet land, including surface matrix}

\item{grooves}{a list object containing the output of the get_grooves_hough function. Should include a left and right groove fit equations}
}
\value{
an x3p file with the added mask that estimates the GEAs from the Hough method.
}
\description{
Function that creates a mask layer for an x3p of a bullet land based on the predictions
of the Hough process.
}
\examples{
library(x3ptools)
data("br411")
x3p <- br411
grooves <- get_grooves_hough(x3p_to_df(x3p), qu = 0.995)

a <- get_mask_hough(x3p, grooves)
}
