% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_hough.R
\name{groove_fit}
\alias{groove_fit}
\title{Calculate groove location from cross cut given line hough line information
Slope and intercept based on the top most and bottom most point of the GEAs.}
\usage{
groove_fit(fit.df, y.input, return_plot = F)
}
\arguments{
\item{y.input}{a numeric y location of where the cross cut was taken from the bullet land}

\item{return_plot}{return plot of grooves as well as location of right and left grooves}

\item{slope}{a numeric slope corresponding to the line determined from the Hough transform}

\item{intercept}{a numeric yintercept corresponding to the line determined from the Hough transform}
}
\value{
a numeric estimate of a groove location based on the equation given from last hough output
}
\description{
Calculate groove location from cross cut given line hough line information
Slope and intercept based on the top most and bottom most point of the GEAs.
}
