% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooves_plot.R
\name{grooves_plot}
\alias{grooves_plot}
\title{Internal function to plot crosscut + grooves}
\usage{
grooves_plot(land, grooves)
}
\arguments{
\item{land}{data.frame with columns x and value}

\item{grooves}{numeric vector of length 2 identifying both grooves.
If only one groove is identified, the other should be NA}
}
\value{
a ggplot2 object
}
\description{
Internal function to plot crosscut + grooves
}
