% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hamby44demo}
\alias{hamby44demo}
\title{hamby44demo}
\format{a list containing lists of links for the 6 lands of bullet 1 and 2}
\usage{
hamby44demo
}
\description{
Links to barrel 1, bullets 1 and 2, with all 6 lands.
}
\keyword{datasets}
