% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_hough.R
\name{pix_to_micron}
\alias{pix_to_micron}
\title{Convert pixel indices to microns}
\usage{
pix_to_micron(x, land)
}
\arguments{
\item{x}{index of pixel from bullet land}

\item{land}{x3p file containing the bullet land}
}
\value{
single numeric of micron location from pixel index
}
\description{
Helper function that should not be used outside of the package. Obtains the
resolution of and x3p-scan, and converts an inputted pixel index to a micron location.
}
