% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_hough.R
\name{rho_to_ab}
\alias{rho_to_ab}
\title{Re-parameterize the Hough transform output for pixel identification}
\usage{
rho_to_ab(rho = NULL, theta = NULL, df = NULL)
}
\arguments{
\item{rho}{Numeric vector containing the shortest distance from the line to the origin}

\item{theta}{Numeric vector containing the angle of the line from the positive x axis}

\item{df}{Data frame containing output from a Hough transformation}
}
\value{
HH: Charlotte, could include some more description about the resulting ouput?
}
\description{
Helper function. Should not be used outside the package.
}
