% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_logistic.R
\name{robust_loess_fit}
\alias{robust_loess_fit}
\title{Fit a robust loess regression}
\usage{
robust_loess_fit(cc, iter)
}
\arguments{
\item{cc}{data frame with columns x and value_std, representing the crosscut}

\item{iter}{number of iterations}
}
\description{
Internal function called by get_grooves_lassobasic and get_grooves_lassofull
}
