% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_hough.R
\name{safely_get_grooves_hough}
\alias{safely_get_grooves_hough}
\title{Return a list of errors and results from get_groove_hough}
\usage{
safely_get_grooves_hough(...)
}
\arguments{
\item{land}{dataframe of surface measurements in microns in the x, y, and x direction}

\item{return_plot}{return plot of grooves}
}
\description{
Return a list of errors and results from get_groove_hough
}
