% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_hough.R
\name{get_grooves_hough}
\alias{get_grooves_hough}
\title{Use Hough transformation to identify groove locations.}
\usage{
get_grooves_hough(land, qu = 0.999, adjust = 10, return_plot = FALSE)
}
\arguments{
\item{land}{dataframe of surface measurements in microns in the x, y, and x direction}

\item{qu}{quantile (between 0 and 1) to specify score quantile for which vertical lines are considered. If groove are not strongly expressed, lower this threshold.}

\item{adjust}{positive number to adjust the grooves inward}

\item{return_plot}{boolean value - should a plot of the crosscut with the grooves be returned? defaults to FALSE}
}
\value{
list object consisting of functions to describe the left and right groove. Parameters for the functions are given in microns and accept micron inputs.
}
\description{
Requires access to full x3p of bullet land for computation. Generates Hough lines in polar coordinate form,
and selects lines with angle from the positive x-axis within the range of negative pi/16 to pi/16
}
\examples{
library(x3ptools)
library(ggplot2)
data("br411")
x3p <- br411

# Get grooves fit for left and right groove from x3p
grooves <- get_grooves_hough(x3p_to_df(x3p), qu = 0.999)

# Find optimized crosscut location, this may take some time
if (require(bulletxtrctr)) {
  crosscut <- x3p \%>\% bulletxtrctr::x3p_crosscut_optimize()
} else {
  crosscut <- 125
}

\dontrun{
a <- get_mask_hough(x3p, grooves)
a <- a \%>\% x3p_add_hline(yintercept=crosscut)
x3ptools::image_x3p(a)
}

# Find groove locations for optimized crosscut
left_groove_hough <- grooves$left.groove.fit(crosscut)
right_groove_hough <- grooves$right.groove.fit(crosscut)

# Plot profile
ccdata <- x3p \%>\% x3p_to_df() \%>\% dplyr::filter(y == crosscut)

ccdata \%>\%
  ggplot(aes(x = x, y = value))+
  geom_line()+
  geom_vline(xintercept = left_groove_hough) +
  geom_vline(xintercept = right_groove_hough)

}
