% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_hough.R
\name{get_grooves_hough}
\alias{get_grooves_hough}
\title{Use Hough transformation to identify groove locations.
Choose strong edges based on whether scores exceed the 99.75% percentile of scores and if the angle of line is less than
Pi/4.}
\usage{
get_grooves_hough(land, return_plot = F)
}
\arguments{
\item{land}{dataframe of surface measurements in microns in the x, y, and x direction}

\item{return_plot}{return plot of grooves}
}
\description{
Use Hough transformation to identify groove locations.
Choose strong edges based on whether scores exceed the 99.75% percentile of scores and if the angle of line is less than
Pi/4.
}
