% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_hough.R
\name{get_grooves_hough}
\alias{get_grooves_hough}
\title{Use Hough transformation to identify groove locations.}
\usage{
get_grooves_hough(land, qu = 0.999, adjust = 10, return_plot = F)
}
\arguments{
\item{land}{dataframe of surface measurements in microns in the x, y, and x direction}

\item{qu}{quantile (between 0 and 1) to specify score quantile for which vertical lines are considered. If groove are not strongly expressed, lower this threshold.}

\item{adjust}{positive number to adjust the grooves inward}

\item{return_plot}{Should in theory return a plot, not available for this method, including for compatability with cc_locate_grooves}
}
\value{
list object consisting of functions to describe the left and right groove
}
\description{
Choose strong edges based on whether scores exceed the 99.75 percentile of scores and if the angle of line is less than
Pi/4.
}
