% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grooves_hough.R
\name{rho_to_ab}
\alias{rho_to_ab}
\title{Re-parameterize the Hough transform output for pixel identification}
\usage{
rho_to_ab(rho = NULL, theta = NULL, df = NULL)
}
\arguments{
\item{rho}{Numeric vector containing the shortest distance from the line to the origin}

\item{theta}{Numeric vector containing the angle of the line from the positive x axis}

\item{df}{Data frame containing output from a Hough transformation}
}
\value{
data frame with variables rho, theta, score (original data frame) expanded by yintercept and xintercept.
}
\description{
Helper function, that should not be used outside the package.
Computes x and y intercepts for a line given in polar coordinate representation.
A line in polar coordinate representation is given by angle theta (with respect to the x-axis) and
distance of the line to the origin rho.
}
